+++ libyafuse2/libfat/dir.c
@@ -118,7 +118,7 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
 
   sz = sizeof(struct msdos_dir_entry);
 
-  ret = io_read((uint8_t *)&dentry, sz);
+  ret = io_read((uint8_t *)&dentry, (size_t)sz);
   if (ret != 0) {
     return -1;
   }
@@ -127,7 +127,7 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
    * Long File Names (LFN), i.e., dslot, is used
    */
   if (dentry.attr == ATTR_EXT) {
-    ret = io_read((uint8_t *)&dentry, sz);
+    ret = io_read((uint8_t *)&dentry, (size_t)sz);
     if (ret != 0) {
       return -1;
     }
@@ -136,7 +136,7 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
   while (dentry.name[0] != '\0') {
     ++i;
 
-    ret = io_read((uint8_t *)&dentry, sz);
+    ret = io_read((uint8_t *)&dentry, (size_t)sz);
     if (ret != 0) {
       return -1;
     }
@@ -145,7 +145,7 @@ int32_t fat_fill_root_dentries(const struct fat_super_block *sb, int32_t *dentri
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dentry.name[0] != '\0' && dentry.attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry, sz);
+      ret = io_read((uint8_t *)&dentry, (size_t)sz);
       if (ret != 0) {
         return -1;
       }
@@ -180,7 +180,7 @@ int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries,
   sz = sizeof(struct msdos_dir_entry);
 
   for (i = 0; i < dentries; ++i) {
-    ret = io_read((uint8_t *)&dslot[i], sz);
+    ret = io_read((uint8_t *)&dslot[i], (size_t)sz);
     if (ret != 0) {
       break;
     }
@@ -189,7 +189,7 @@ int32_t fat_fill_root_dentry(const struct fat_super_block *sb, int32_t dentries,
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dslot[i].attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry[i], sz);
+      ret = io_read((uint8_t *)&dentry[i], (size_t)sz);
       if (ret != 0) {
         break;
       }
@@ -227,7 +227,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
   /*
    * Read FAT dentry of '.'
    */
-  ret = io_read((uint8_t *)&dentry, sz);
+  ret = io_read((uint8_t *)&dentry, (size_t)sz);
   if (ret != 0) {
     return -1;
   }
@@ -235,7 +235,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
   /*
    * Read FAT dentry of '..'
    */
-  ret = io_read((uint8_t *)&dentry, sz);
+  ret = io_read((uint8_t *)&dentry, (size_t)sz);
   if (ret != 0) {
     return -1;
   }
@@ -243,7 +243,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
   while (dentry.name[0] != '\0') {
     ++i;
 
-    ret = io_read((uint8_t *)&dentry, sz);
+    ret = io_read((uint8_t *)&dentry, (size_t)sz);
     if (ret != 0) {
       return -1;
     }
@@ -252,7 +252,7 @@ int32_t fat_fill_dentries(const struct fat_super_block *sb, int32_t cluster, int
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dentry.attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry, sz);
+      ret = io_read((uint8_t *)&dentry, (size_t)sz);
       if (ret != 0) {
         return -1;
       }
@@ -291,7 +291,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
    */
   memset((void *)&dslot[0], 0, sz);
 
-  ret = io_read((uint8_t *)&dentry[0], sz);
+  ret = io_read((uint8_t *)&dentry[0], (size_t)sz);
   if (ret != 0) {
     return -1;
   }
@@ -301,7 +301,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
    */
   memset((void *)&dslot[1], 0, sz);
 
-  ret = io_read((uint8_t *)&dentry[1], sz);
+  ret = io_read((uint8_t *)&dentry[1], (size_t)sz);
   if (ret != 0) {
     return -1;
   }
@@ -310,7 +310,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
    * Fill in other dentry
    */
   for (i = 2; i < dentries; ++i) {
-    ret = io_read((uint8_t *)&dslot[i], sz);
+    ret = io_read((uint8_t *)&dslot[i], (size_t)sz);
     if (ret != 0) {
       break;
     }
@@ -319,7 +319,7 @@ int32_t fat_fill_dentry(const struct fat_super_block *sb, int32_t cluster, int32
      * Long File Names (LFN), i.e., dslot, is used
      */
     if (dslot[i].attr == ATTR_EXT) {
-      ret = io_read((uint8_t *)&dentry[i], sz);
+      ret = io_read((uint8_t *)&dentry[i], (size_t)sz);
       if (ret != 0) {
         break;
       }
